refactor: 멤버 관련 이벤트 처리 로직 멤버 핸들러로 이동
